Move ToggleYesNo into it's own directory.
