+++ src/index.js
@@ -8,7 +8,8 @@ import LoaderContainer from './components/LoaderContainer';
 import ProgressNavigation from './components/ProgressNavigation';
 import Select from './components/Select';
 import Tabs from './components/Tabs';
-import {Toggle, ToggleYesNo} from './components/Toggle';
+import Toggle from './components/Toggle';
+import ToggleYesNo from './components/ToggleYesNo';
 import USP from './components/USP';
 import ValidationError from './components/ValidationError';
 import ValidationMessage from './components/ValidationMessage';
